Update Dockerfile
Hi, in [this learning lab](https://developer.cisco.com/learning/labs/app-dynamics-docker/download-and-build-source-code/), running docker-compose build in this directory result sin an error:

 => [2/6] RUN apt-get update                                                               7.1s
 => ERROR [3/6] RUN apt-get install -y yum                                                 1.2s
------
 > [3/6] RUN apt-get install -y yum:
#0 0.264 Reading package lists...
#0 1.025 Building dependency tree...
#0 1.185 Reading state information...
#0 1.193 E: Unable to locate package yum

The error message is telling us that the package yum cannot be found in the package repository that apt-get is trying to search. This is because yum is a package manager used by Red Hat based systems, while apt-get is used by Debian based systems.

Since the Tomcat Docker image is based on a Debian system, it uses apt-get as its package manager. Therefore, yum cannot be installed using apt-get.

I've tested this change and it works.